agentic_ai/agno_yfinance/requirements.txt
agno
anthropic
duckduckgo-search
yfinance
python-dotenv
ddgs
